@@ -46,11 +46,9 @@
     "source": "https://github.com/WikibaseSolutions/php-cypher-dsl"
   },
   "require": {
-    "php": ">=7.4",
+    "php": ">=8.1",
     "ext-ctype": "*",
-    "ext-openssl": "*",
-    "symfony/polyfill-php80": "^1.25",
-    "symfony/polyfill-php81": "^1.25"
+    "ext-openssl": "*"
   },
   "require-dev": {
     "phpunit/phpunit": "~9.0",
