@@ -11,8 +11,8 @@ namespace WikibaseSolutions\CypherDSL\Clauses;
 
 use WikibaseSolutions\CypherDSL\Patterns\Pattern;
 use WikibaseSolutions\CypherDSL\Query;
-use WikibaseSolutions\CypherDSL\Traits\CastTrait;
 use WikibaseSolutions\CypherDSL\Types\StructuralTypes\StructuralType;
+use WikibaseSolutions\CypherDSL\Utils\CastUtils;
 
 /**
  * This class represents a DELETE clause.
@@ -25,8 +25,6 @@ use WikibaseSolutions\CypherDSL\Types\StructuralTypes\StructuralType;
  */
 final class DeleteClause extends Clause
 {
-    use CastTrait;
-
     /**
      * Whether the DETACH modifier is needed.
      */
@@ -64,7 +62,7 @@ final class DeleteClause extends Clause
         $res = [];
 
         foreach ($structures as $structure) {
-            $res[] = self::toStructuralType($structure);
+            $res[] = CastUtils::toStructuralType($structure);
         }
 
         $this->structures = array_merge($this->structures, $res);
