@@ -10,6 +10,7 @@
 namespace WikibaseSolutions\CypherDSL\Expressions\Literals;
 
 use LogicException;
+use Stringable;
 use Traversable;
 use TypeError;
 use WikibaseSolutions\CypherDSL\Expressions\Procedures\Date;
@@ -19,19 +20,16 @@ use WikibaseSolutions\CypherDSL\Expressions\Procedures\LocalTime;
 use WikibaseSolutions\CypherDSL\Expressions\Procedures\Point;
 use WikibaseSolutions\CypherDSL\Expressions\Procedures\Procedure;
 use WikibaseSolutions\CypherDSL\Expressions\Procedures\Time;
-use WikibaseSolutions\CypherDSL\Traits\ErrorTrait;
 use WikibaseSolutions\CypherDSL\Types\PropertyTypes\NumeralType;
 use WikibaseSolutions\CypherDSL\Types\PropertyTypes\StringType;
 
 /**
  * Helper class to construct literals.
  *
- * @note This class should only contain static methods
+ * @note This class should only contain static methods.
  */
 final class Literal
 {
-    use ErrorTrait;
-
     /**
      * Creates a new literal from the given value. This function automatically constructs the appropriate
      * class based on the type of the value given.
@@ -42,7 +40,7 @@ final class Literal
      *
      * @return Boolean|Float_|Integer|List_|Map|String_
      */
-    public static function literal($literal)
+    public static function literal(string|bool|int|float|array|Stringable $literal): Boolean|Float_|Integer|List_|Map|String_
     {
         if (is_string($literal)) {
             return self::string($literal);
@@ -82,7 +80,7 @@ final class Literal
      * @see        Literal::number()
      * @deprecated
      */
-    public static function decimal($value)
+    public static function decimal(float|int $value): Float_|Integer
     {
         return self::number($value);
     }
@@ -94,10 +92,8 @@ final class Literal
      *
      * @return Float_|Integer
      */
-    public static function number($value)
+    public static function number(float|int $value): Float_|Integer
     {
-        self::assertClass('value', ['int', 'float'], $value);
-
         // @phpstan-ignore-next-line PHPStan does not work well with classes named "Integer"
         return is_int($value) ? self::integer($value) : self::float($value);
     }
@@ -155,7 +151,7 @@ final class Literal
     /**
      * Creates a new map literal.
      *
-     * @param mixed[] $value
+     * @param array $value
      */
     public static function map(array $value): Map
     {
@@ -168,9 +164,11 @@ final class Literal
      *
      * @param null|string|StringType $timezone
      *
+     * @return Date
+     *
      * @see https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-date-current Corresponding documentation on Neo4j.com
      */
-    public static function date($timezone = null): Date
+    public static function date(StringType|string|null $timezone = null): Date
     {
         if ($timezone === null) {
             return Procedure::date();
@@ -182,13 +180,15 @@ final class Literal
     /**
      * Creates a date from the given year, month and day.
      *
-     * @param int|NumeralType      $year
+     * @param int|NumeralType $year
      * @param null|int|NumeralType $month
      * @param null|int|NumeralType $day
      *
+     * @return Date
+     *
      * @see https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-date-calendar Corresponding documentation on Neo4j.com
      */
-    public static function dateYMD($year, $month = null, $day = null): Date
+    public static function dateYMD(NumeralType|int $year, NumeralType|int|null $month = null, NumeralType|int|null $day = null): Date
     {
         return Procedure::date(
             self::makeTemporalMap(
@@ -204,13 +204,15 @@ final class Literal
     /**
      * Creates a date from the given year, week and weekday.
      *
-     * @param int|NumeralType      $year
+     * @param int|NumeralType $year
      * @param null|int|NumeralType $week
      * @param null|int|NumeralType $weekday
      *
+     * @return Date
+     *
      * @see https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-date-week Corresponding documentation on Neo4j.com
      */
-    public static function dateYWD($year, $week = null, $weekday = null): Date
+    public static function dateYWD(NumeralType|int $year, NumeralType|int|null $week = null, NumeralType|int|null $weekday = null): Date
     {
         return Procedure::date(
             self::makeTemporalMap(
@@ -228,9 +230,11 @@ final class Literal
      *
      * @param string|StringType $date
      *
+     * @return Date
+     *
      * @see https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-date-create-string Corresponding documentation on Neo4j.com
      */
-    public static function dateString($date): Date
+    public static function dateString(StringType|string $date): Date
     {
         return Procedure::date($date);
     }
@@ -239,11 +243,13 @@ final class Literal
      * Retrieves the current DateTime value, optionally for a different time zone. In reality, this
      * function just returns a call to the "datetime()" function.
      *
-     * @param string|StringType $timezone
+     * @param StringType|string|null $timezone
+     *
+     * @return DateTime
      *
      * @see https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-datetime-current Corresponding documentation on Neo4j.com
      */
-    public static function dateTime($timezone = null): DateTime
+    public static function dateTime(StringType|string|null $timezone = null): DateTime
     {
         if ($timezone === null) {
             return Procedure::datetime();
@@ -269,16 +275,16 @@ final class Literal
      * @see https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-datetime-calendar Corresponding documentation on Neo4j.com
      */
     public static function dateTimeYMD(
-        $year,
-        $month = null,
-        $day = null,
-        $hour = null,
-        $minute = null,
-        $second = null,
-        $millisecond = null,
-        $microsecond = null,
-        $nanosecond = null,
-        $timezone = null
+        NumeralType|int $year,
+        NumeralType|int|null $month = null,
+        NumeralType|int|null $day = null,
+        NumeralType|int|null $hour = null,
+        NumeralType|int|null $minute = null,
+        NumeralType|int|null $second = null,
+        NumeralType|int|null $millisecond = null,
+        NumeralType|int|null $microsecond = null,
+        NumeralType|int|null $nanosecond = null,
+        StringType|string|null $timezone = null
     ): DateTime {
         return Procedure::datetime(
             self::makeTemporalMap(
@@ -315,16 +321,16 @@ final class Literal
      * @see https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-datetime-week Corresponding documentation on Neo4j.com
      */
     public static function dateTimeYWD(
-        $year,
-        $week = null,
-        $dayOfWeek = null,
-        $hour = null,
-        $minute = null,
-        $second = null,
-        $millisecond = null,
-        $microsecond = null,
-        $nanosecond = null,
-        $timezone = null
+        NumeralType|int $year,
+        NumeralType|int|null $week = null,
+        NumeralType|int|null $dayOfWeek = null,
+        NumeralType|int|null $hour = null,
+        NumeralType|int|null $minute = null,
+        NumeralType|int|null $second = null,
+        NumeralType|int|null $millisecond = null,
+        NumeralType|int|null $microsecond = null,
+        NumeralType|int|null $nanosecond = null,
+        StringType|string|null $timezone = null
     ): DateTime {
         return Procedure::datetime(
             self::makeTemporalMap(
@@ -361,16 +367,16 @@ final class Literal
      * @see https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-datetime-quarter Corresponding documentation on Neo4j.com
      */
     public static function dateTimeYQD(
-        $year,
-        $quarter = null,
-        $dayOfQuarter = null,
-        $hour = null,
-        $minute = null,
-        $second = null,
-        $millisecond = null,
-        $microsecond = null,
-        $nanosecond = null,
-        $timezone = null
+        NumeralType|int $year,
+        NumeralType|int|null $quarter = null,
+        NumeralType|int|null $dayOfQuarter = null,
+        NumeralType|int|null $hour = null,
+        NumeralType|int|null $minute = null,
+        NumeralType|int|null $second = null,
+        NumeralType|int|null $millisecond = null,
+        NumeralType|int|null $microsecond = null,
+        NumeralType|int|null $nanosecond = null,
+        StringType|string|null $timezone = null
     ): DateTime {
         return Procedure::datetime(
             self::makeTemporalMap(
@@ -406,15 +412,15 @@ final class Literal
      * @see https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-datetime-ordinal Corresponding documentation on Neo4j.com
      */
     public static function dateTimeYD(
-        $year,
-        $ordinalDay = null,
-        $hour = null,
-        $minute = null,
-        $second = null,
-        $millisecond = null,
-        $microsecond = null,
-        $nanosecond = null,
-        $timezone = null
+        NumeralType|int $year,
+        NumeralType|int|null $ordinalDay = null,
+        NumeralType|int|null $hour = null,
+        NumeralType|int|null $minute = null,
+        NumeralType|int|null $second = null,
+        NumeralType|int|null $millisecond = null,
+        NumeralType|int|null $microsecond = null,
+        NumeralType|int|null $nanosecond = null,
+        StringType|string|null $timezone = null
     ): DateTime {
         return Procedure::datetime(
             self::makeTemporalMap(
@@ -438,7 +444,7 @@ final class Literal
      *
      * @param string|StringType $dateString
      */
-    public static function dateTimeString($dateString): DateTime
+    public static function dateTimeString(StringType|string $dateString): DateTime
     {
         return Procedure::datetime($dateString);
     }
@@ -450,7 +456,7 @@ final class Literal
      *
      * @see https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-localdatetime-current Corresponding documentation on Neo4j.com
      */
-    public static function localDateTime($timezone = null): LocalDateTime
+    public static function localDateTime(StringType|string|null $timezone = null): LocalDateTime
     {
         if ($timezone === null) {
             return Procedure::localdatetime();
@@ -475,15 +481,15 @@ final class Literal
      * @see https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-localdatetime-calendar Corresponding documentation on Neo4j.com
      */
     public static function localDateTimeYMD(
-        $year,
-        $month = null,
-        $day = null,
-        $hour = null,
-        $minute = null,
-        $second = null,
-        $millisecond = null,
-        $microsecond = null,
-        $nanosecond = null
+        NumeralType|int $year,
+        NumeralType|int|null $month = null,
+        NumeralType|int|null $day = null,
+        NumeralType|int|null $hour = null,
+        NumeralType|int|null $minute = null,
+        NumeralType|int|null $second = null,
+        NumeralType|int|null $millisecond = null,
+        NumeralType|int|null $microsecond = null,
+        NumeralType|int|null $nanosecond = null
     ): LocalDateTime {
         return Procedure::localdatetime(
             self::makeTemporalMap(
@@ -506,7 +512,7 @@ final class Literal
      * Creates a LocalDateTime value with the specified year, week, dayOfWeek, hour, minute,
      * second, millisecond, microsecond and nanosecond component value.
      *
-     * @param int|NumeralType      $year
+     * @param int|NumeralType $year
      * @param null|int|NumeralType $week
      * @param null|int|NumeralType $dayOfWeek
      * @param null|int|NumeralType $hour
@@ -516,18 +522,20 @@ final class Literal
      * @param null|int|NumeralType $microsecond
      * @param null|int|NumeralType $nanosecond
      *
+     * @return LocalDateTime
+     *
      * @see https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-localdatetime-week Corresponding documentation on Neo4j.com
      */
     public static function localDateTimeYWD(
-        $year,
-        $week = null,
-        $dayOfWeek = null,
-        $hour = null,
-        $minute = null,
-        $second = null,
-        $millisecond = null,
-        $microsecond = null,
-        $nanosecond = null
+        NumeralType|int $year,
+        NumeralType|int|null $week = null,
+        NumeralType|int|null $dayOfWeek = null,
+        NumeralType|int|null $hour = null,
+        NumeralType|int|null $minute = null,
+        NumeralType|int|null $second = null,
+        NumeralType|int|null $millisecond = null,
+        NumeralType|int|null $microsecond = null,
+        NumeralType|int|null $nanosecond = null
     ): LocalDateTime {
         return Procedure::localdatetime(
             self::makeTemporalMap(
@@ -562,15 +570,15 @@ final class Literal
      * @see https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-localdatetime-quarter Corresponding documentation on Neo4j.com
      */
     public static function localDateTimeYQD(
-        $year,
-        $quarter = null,
-        $dayOfQuarter = null,
-        $hour = null,
-        $minute = null,
-        $second = null,
-        $millisecond = null,
-        $microsecond = null,
-        $nanosecond = null
+        NumeralType|int $year,
+        NumeralType|int|null $quarter = null,
+        NumeralType|int|null $dayOfQuarter = null,
+        NumeralType|int|null $hour = null,
+        NumeralType|int|null $minute = null,
+        NumeralType|int|null $second = null,
+        NumeralType|int|null $millisecond = null,
+        NumeralType|int|null $microsecond = null,
+        NumeralType|int|null $nanosecond = null
     ): LocalDateTime {
         return Procedure::localdatetime(
             self::MakeTemporalMap(
@@ -604,14 +612,14 @@ final class Literal
      * @see https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-localdatetime-ordinal Corresponding documentation on Neo4j.com
      */
     public static function localDateTimeYD(
-        $year,
-        $ordinalDay = null,
-        $hour = null,
-        $minute = null,
-        $second = null,
-        $millisecond = null,
-        $microsecond = null,
-        $nanosecond = null
+        NumeralType|int $year,
+        NumeralType|int|null $ordinalDay = null,
+        NumeralType|int|null $hour = null,
+        NumeralType|int|null $minute = null,
+        NumeralType|int|null $second = null,
+        NumeralType|int|null $millisecond = null,
+        NumeralType|int|null $microsecond = null,
+        NumeralType|int|null $nanosecond = null
     ): LocalDateTime {
         return Procedure::localdatetime(
             self::makeTemporalMap(
@@ -636,7 +644,7 @@ final class Literal
      *
      * @see https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-localdatetime-create-string Corresponding documentation on Neo4j.com
      */
-    public static function localDateTimeString($localDateTimeString): LocalDateTime
+    public static function localDateTimeString(StringType|string $localDateTimeString): LocalDateTime
     {
         return Procedure::localdatetime($localDateTimeString);
     }
@@ -648,7 +656,7 @@ final class Literal
      *
      * @see https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-localtime-current Corresponding documentation on Neo4j.com
      */
-    public static function localTimeCurrent($timezone = null): LocalTime
+    public static function localTimeCurrent(StringType|string|null $timezone = null): LocalTime
     {
         if ($timezone === null) {
             return Procedure::localtime();
@@ -670,12 +678,12 @@ final class Literal
      * @see https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-localtime-create Corresponding documentation on Neo4j.com
      */
     public static function localTime(
-        $hour,
-        $minute = null,
-        $second = null,
-        $millisecond = null,
-        $microsecond = null,
-        $nanosecond = null
+        NumeralType|int $hour,
+        NumeralType|int|null $minute = null,
+        NumeralType|int|null $second = null,
+        NumeralType|int|null $millisecond = null,
+        NumeralType|int|null $microsecond = null,
+        NumeralType|int|null $nanosecond = null
     ): LocalTime {
         return Procedure::localtime(
             self::makeTemporalMap(
@@ -696,7 +704,7 @@ final class Literal
      *
      * @param string|StringType $localTimeString
      */
-    public static function localTimeString($localTimeString): LocalTime
+    public static function localTimeString(StringType|string $localTimeString): LocalTime
     {
         return Procedure::localtime($localTimeString);
     }
@@ -708,7 +716,7 @@ final class Literal
      *
      * @see https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-time-current Corresponding documentation on Neo4j.com
      */
-    public static function time($timezone = null): Time
+    public static function time(StringType|string|null $timezone = null): Time
     {
         if ($timezone === null) {
             return Procedure::time();
@@ -731,13 +739,13 @@ final class Literal
      * @see https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-time-create Corresponding documentation on Neo4j.com
      */
     public static function timeHMS(
-        $hour,
-        $minute = null,
-        $second = null,
-        $millisecond = null,
-        $microsecond = null,
-        $nanosecond = null,
-        $timezone = null
+        NumeralType|int $hour,
+        NumeralType|int|null $minute = null,
+        NumeralType|int|null $second = null,
+        NumeralType|int|null $millisecond = null,
+        NumeralType|int|null $microsecond = null,
+        NumeralType|int|null $nanosecond = null,
+        NumeralType|int|null $timezone = null
     ): Time {
         return Procedure::time(
             self::makeTemporalMap(
@@ -761,7 +769,7 @@ final class Literal
      *
      * @see https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-time-create-string Corresponding documentation on Neo4j.com
      */
-    public static function timeString($timeString): Time
+    public static function timeString(StringType|string $timeString): Time
     {
         return Procedure::time($timeString);
     }
@@ -774,7 +782,7 @@ final class Literal
      *
      * @see https://neo4j.com/docs/cypher-manual/current/functions/spatial/#functions-point-cartesian-2d Corresponding documentation on Neo4j.com
      */
-    public static function point2d($x, $y): Point
+    public static function point2d(NumeralType|int|float $x, NumeralType|int|float $y): Point
     {
         return Procedure::point([
             "x" => $x,
@@ -792,7 +800,7 @@ final class Literal
      *
      * @see https://neo4j.com/docs/cypher-manual/current/functions/spatial/#functions-point-cartesian-3d Corresponding documentation on Neo4j.com
      */
-    public static function point3d($x, $y, $z): Point
+    public static function point3d(NumeralType|int|float $x, NumeralType|int|float $y, NumeralType|int|float $z): Point
     {
         return Procedure::point([
             "x" => $x,
@@ -810,7 +818,7 @@ final class Literal
      *
      * @see https://neo4j.com/docs/cypher-manual/current/functions/spatial/#functions-point-wgs84-2d Corresponding documentation on Neo4j.com
      */
-    public static function point2dWGS84($longitude, $latitude): Point
+    public static function point2dWGS84(NumeralType|int|float $longitude, NumeralType|int|float $latitude): Point
     {
         return Procedure::point([
             "longitude" => $longitude,
@@ -828,7 +836,7 @@ final class Literal
      *
      * @see https://neo4j.com/docs/cypher-manual/current/functions/spatial/#functions-point-wgs84-2d Corresponding documentation on Neo4j.com
      */
-    public static function point3dWGS84($longitude, $latitude, $height): Point
+    public static function point3dWGS84(NumeralType|int|float $longitude, NumeralType|int|float $latitude, NumeralType|int|float $height): Point
     {
         return Procedure::point([
             "longitude" => $longitude,
@@ -848,9 +856,9 @@ final class Literal
     /**
      * Prepares the variables to be used by temporal (i.e. time-like) CYPHER-functions.
      *
-     * @param mixed[] $variables
+     * @param array $variables
      *
-     * @return mixed[]
+     * @return array
      */
     private static function makeTemporalMap(array $variables): array
     {
